+++ includes/traits/trait-admin-page-sites.php
@@ -19,5 +19,6 @@ trait Alynt_Drime_Backups_Dashboard_Admin_Page_Sites {
 	use Alynt_Drime_Backups_Dashboard_Admin_Page_Sites_List;
 	use Alynt_Drime_Backups_Dashboard_Admin_Page_Add_Site;
 	use Alynt_Drime_Backups_Dashboard_Admin_Page_Attention;
+	use Alynt_Drime_Backups_Dashboard_Admin_Page_Site_Detail_Local_Record_Panels;
 	use Alynt_Drime_Backups_Dashboard_Admin_Page_Site_Detail;
 }